// src/app/pages/configuracao/models/setupWeb.model.ts
export interface SetupWebModel {
    idSetupWeb?: number;
    idConfiguracao: number;
    nomeDoSite: string;
    enderecoDoSite: string;
    imagemDoIcone: string;
    textoMensagemPermissao: string;
    textoBotaoPermitir: string;
    textoBotaoNegar: string;
    tituloNotificacaoBoasVindas: string;
    textoMensagemBoasVindas: string;
    habilitarLinkDestino: boolean;
    enderecoLinkDestino: string;
  }
  